+++ app/src/main/java/com/nkraft/eyebox/services/TermsService.java
@@ -27,7 +27,7 @@ public class TermsService extends BaseService {
             List<Terms> termsList = new ArrayList<>();
             for (int i = 0; i < jsonArray.length(); ++i) {
                 JSONObject jsonObject = jsonArray.getJSONObject(i);
-                long id = jsonObject.getLong("id");
+                int id = jsonObject.getInt("id");
                 int days = jsonObject.getInt("days");
                 String namex = jsonObject.getString("namex");
                 int dateDelay = jsonObject.getInt("datedelay");
